+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated logic for displaying multiple zeros on Y and Y2 axis ticks.
+  
 ## 2.19.0 - (November 9, 2021)
 
 * Changed